+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/CustomerTablePanel.java
@@ -50,7 +50,7 @@ public final class CustomerTablePanel extends EntityTablePanel {
 	}
 
 	@Override
-	protected void configureControls() {
+	protected void setupControls() {
 		control(PRINT).set(Control.builder(this::viewCustomerReport)
 						.name(BUNDLE.getString("customer_report"))
 						.smallIcon(FrameworkIcons.instance().print())